Add tertiary colours for user-selection and extract colors to Colors.kt. Minor changes to the ColorSelectionDialog. If colors are not found in the database when assigning to a genus, then the repo will create the color entity.

+++ app/src/main/java/com/bp/dinodata/db/entities/GenusColorUpdate.kt
@@ -6,5 +6,5 @@ class GenusColorUpdate(
     @ColumnInfo(name = "name")
     val genusName: String,
     @ColumnInfo(name = "color_id")
-    val colorId: Int? = null
+    val colorId: Long? = null
 )
